requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
cryptography==42.0.2
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.101.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
idna==3.4
nest-asyncio==1.6.0
passlib==1.7.4
psycopg2==2.9.7
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
strawberry-graphql==0.203.0
typing_extensions==4.7.1
urllib3==2.2.0
uvicorn==0.23.2